% Generated by roxygen2 (4.0.2): do not edit by hand
\name{compute_shares_remain}
\alias{compute_shares_remain}
\title{Compute the Share of Remaining Participants}
\usage{
compute_shares_remain(df, drop_out_pos, number_of_questions)
}
\arguments{
\item{df}{a data.frame}

\item{drop_out_pos}{drop out position}

\item{number_of_questions}{integer number of questions}
}
\description{
Compute the share of remaining participant per question.
}

