% Generated by roxygen2 (4.0.2): do not edit by hand
\name{compute_xsq}
\alias{compute_xsq}
\title{Compute Chisq test for a list}
\usage{
compute_xsq(li, pos, participants, sel)
}
\arguments{
\item{li}{a list}

\item{pos}{integer position within the respective data.frane}

\item{participants}{known integer}

\item{sel}{integer position in th result table}
}
\description{
Compute chisq given a list of data.frames
}

