% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeRemaining.R
\name{computeRemaining}
\alias{computeRemaining}
\title{Compute Share of Remaining Participants}
\usage{
computeRemaining(dframe, n_cols, by_cond = NULL, long_format = T,
  d_col = "drop_out", strip_last_question = T)
}
\arguments{
\item{dframe}{a data.frame}

\item{number}{of questions}
}
\description{
Given a one-row-one-observation, one-column-one-question format data.frame,
which uses R compliant coding (i.e. NA) for missing values,
this function computes the share of remaining participants for each question.
}

