% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeStatistics.R
\name{computeStatistics}
\alias{computeStatistics}
\title{Compute Statistics}
\usage{
computeStatistics(
  df,
  by_cond = "None",
  do_indicator = "drop_out_idx",
  no_of_vars
)
}
\arguments{
\item{df}{data.frame}

\item{by_cond}{character group by condition, defaults to None.}

\item{do_indicator}{character dropout indicator}

\item{no_of_vars}{numeric number of variables}
}
\description{
Compute Statistics
}
