% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_ks.R
\name{do_ks}
\alias{do_ks}
\title{Compute Kolmogorov-Smirnov Test for most extreme conditions}
\usage{
do_ks(data, questions)
}
\arguments{
\item{data}{A data frame made from [compute_stats()], containing information on the percent remaining per question per condition}

\item{questions}{Number of questions to be included in analysis}
}
\value{
result of Kolmogorov-Smirnoff test
}
\description{
This test is used for survival analysis between the most extreme conditions,
so the ones with the most different rates of dropout.
This function automatically prepares your data to run `stats::ks.test()` on it.
}
