% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_or_table.R
\name{do_or_table}
\alias{do_or_table}
\title{Dropout Odds Ratio Table}
\usage{
do_or_table(df, chisq_question, sel_cond_chisq)
}
\arguments{
\item{df}{data.frame stats table as computed by compute_stats.}

\item{chisq_question}{description}

\item{sel_cond_chisq}{description}
}
\description{
Dropout Odds Ratio Table
}
