% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_steps.R
\name{do_steps}
\alias{do_steps}
\title{Calculate Steps for Uneven Data Points}
\usage{
do_steps(x, y, return_df = T)
}
\arguments{
\item{x}{Numeric vector representing the x-coordinates.}

\item{y}{Numeric vector representing the y-coordinates.}

\item{return_df}{Logical. If \code{TRUE}, the function returns a data frame; otherwise, it returns a list.}
}
\value{
A data frame or a list containing the modified \code{x} and \code{y} coordinates.
}
\description{
The \code{do_steps} function calculates steps for uneven data points represented by vectors \code{x} and \code{y}.
}
\examples{
x <- c(1, 2, 3, 4, 5)
y <- c(10, 20, 30, 40, 50)
do_steps(x, y)

}
