% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{dropout}
\alias{dropout}
\title{Dropout in a random dataset}
\format{A data.frame with 1000 observations and 12 variables
\describe{
  \item{qXX}{question}
  \item{group}{experimental condition}
  ...
}}
\source{
A random dataset.
}
\usage{
dropout
}
\description{
A random dataset that mimmicks a survey or experiment with
dropout.
}
\keyword{datasets}

