% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dropRdemo.R
\name{dropRdemo}
\alias{dropRdemo}
\title{Demo Dataset for Dropout in an Online Survey}
\format{
A data frame with 246 rows and 54 variables (in the order they were presented in the fictional survey).
\describe{
\item{obs_id}{Observation ID}
\item{experimental_condition}{experimental condition}
\item{vi_1}{item 1}
\item{vi_2}{item 2}
\item{vi_3}{item 3}
\item{vi_4}{item 4}
\item{vi_5}{item 5}
\item{vi_6}{item 6}
\item{vi_7}{item 7}
\item{vi_8}{item 8}
\item{vi_9}{item 9}
\item{vi_10}{item 10}
\item{vi_11}{item 11}
\item{vi_12}{item 12}
\item{vi_13}{item 13}
\item{vi_14}{item 14}
\item{vi_15}{item 15}
\item{vi_16}{item 16}
\item{vi_17}{item 17}
\item{vi_18}{item 18}
\item{vi_19}{item 19}
\item{vi_20}{item 20}
\item{vi_21}{item 21}
\item{vi_22}{item 22}
\item{vi_23}{item 23}
\item{vi_24}{item 24}
\item{vi_25}{item 25}
\item{vi_26}{item 26}
\item{vi_27}{item 27}
\item{vi_28}{item 28}
\item{vi_29}{item 29}
\item{vi_30}{item 30}
\item{vi_31}{item 31}
\item{vi_32}{item 32}
\item{vi_33}{item 33}
\item{vi_34}{item 34}
\item{vi_35}{item 35}
\item{vi_36}{item 36}
\item{vi_37}{item 37}
\item{vi_38}{item 38}
\item{vi_39}{item 39}
\item{vi_40}{item 40}
\item{vi_41}{item 41}
\item{vi_42}{item 42}
\item{vi_43}{item 43}
\item{vi_44}{item 44}
\item{vi_45}{item 45}
\item{vi_46}{item 46}
\item{vi_47}{item 47}
\item{vi_48}{item 48}
\item{vi_49}{item 49}
\item{vi_50}{item 50}
\item{vi_51}{item 51}
\item{vi_52}{item 52}
}
}
\source{
{dropRdemo} Demo data for dropout.
}
\description{
Simulated demo data set for dropout in a survey.
}
