% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_odds.R
\name{get_odds}
\alias{get_odds}
\title{Compute Odds From Probabilities}
\usage{
get_odds(p)
}
\arguments{
\item{p}{vector of probabilities. May not be larger than 1 or smaller than zero.}
}
\description{
Compute odds from probabilities. The function is vectorized and
can handle a vector of probabilities.
}
\examples{
get_odds(0.7)
get_odds(c(0.7, 0.2))


}
