% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_odds.R
\name{get_odds_ratio}
\alias{get_odds_ratio}
\title{Compute Odds Ratio}
\usage{
get_odds_ratio(a, b)
}
\arguments{
\item{a}{numeric probability value between 0 and 1.}

\item{b}{numeric probability value between 0 and 1.}
}
\description{
Computes odds ratio given two probabilities.
In this package, the function can be used to compare the percentages of remaining
participants between two conditions at a time.
}
\examples{
get_odds_ratio(0.7, 0.6)

}
\seealso{
\code{\link[=get_odds]{get_odds()}}, as this is the basis for calculation.
}
