% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_steps_by_cond}
\alias{get_steps_by_cond}
\title{Get Steps Data by Condition}
\usage{
get_steps_by_cond(sfit, condition = NULL)
}
\arguments{
\item{sfit}{An object representing survival model results (e.g., from a Kaplan-Meier model).}

\item{condition}{Optional. A condition or group label to include in the output data frame.}
}
\value{
A data frame containing the steps data, including time, survival estimates, upper confidence bounds, and lower confidence bounds.
}
\description{
The \code{get_steps_by_cond} function calculates steps data based on survival model results.
}
