% Generated by roxygen2 (4.0.2): do not edit by hand
\name{lineChartOutput}
\alias{lineChartOutput}
\title{Show Line Chart Output on a HTML canvas}
\usage{
lineChartOutput(inputId, width = "100\%", height = "400px")
}
\arguments{
\item{inputId}{input identifier for the output function, i.e. name of the list element in shiny.}

\item{width}{defaults to 100\%.}

\item{height}{defaults to 400px.}
}
\description{
create a HTML SVG Line Chart Output using NVD3.js
This function should be called from ui.R
in a shiny web application.
}

