% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{n_count}
\alias{n_count}
\title{The `n_count` function calculates the number of occurrences of a specific value `n` in a numeric vector `x`.}
\usage{
n_count(x, n)
}
\arguments{
\item{x}{Numeric vector in which to count occurrences.}

\item{n}{Numeric value to count occurrences of.}
}
\value{
An integer representing the count of occurrences of `n` in `x`.
}
\description{
The `n_count` function calculates the number of occurrences of a specific value `n` in a numeric vector `x`.
}
\examples{
x <- c(1, 2, 3, 2, 4, 2)
n_count(x, 2)  # Returns 3

}
