% Generated by roxygen2 (4.0.2): do not edit by hand
\name{renderLineChart}
\alias{renderLineChart}
\title{Render a line Chart in shiny}
\usage{
renderLineChart(expr, env = parent.frame(), quoted = FALSE)
}
\arguments{
\item{expr}{an expression to rendered}

\item{env}{environment, defaults to parent.frame().}

\item{quoted}{logical, defaults to FALSE.}
}
\description{
This function renders a line chart and should be called from server.R in a shiny application.
}

