% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/start_app.R
\name{start_app}
\alias{start_app}
\title{Start the dropR Shiny App}
\usage{
start_app()
}
\description{
Starts the interactive web application to use dropR in your web browser. 
Make sure to use Google Chrome or Firefox for best experience.
}
