% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/startApp.R
\name{startApp}
\alias{startApp}
\title{Start the dropR Shiny App}
\usage{
startApp()
}
\description{
Starts the interactive web application to use dropR in your web browser. 
Make sure to use Google Chrome or Firefox for best experience.
}
