% Generated by roxygen2 (4.0.2): do not edit by hand
\name{startDropR}
\alias{startDropR}
\title{Start the DropR Shiny App}
\usage{
startDropR()
}
\description{
Starts the interactive web application to use dropR in your web browser.
Make sure to use Google Chrome or Firefox for best experience.
}
\examples{
\dontrun{startdropR()}
}

