% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_drop_out_from_df.R
\name{add_dropout_idx}
\alias{add_dropout_idx}
\title{Extract Drop Out from a Data.Frame}
\usage{
add_dropout_idx(df, q_pos)
}
\arguments{
\item{df}{a data.frame}

\item{q_pos}{numeric columns that contain questions}
}
\description{
Find drop out posititions in a data.frame that contains multiple 
questions that had been asked sequentially.
}
