% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{all_values}
\alias{all_values}
\title{Concatenate Names and Values of a Named Vector}
\source{
R/utils.R

The \code{all_values} function takes a named vector \code{x} and concatenates the names
and formatted values into a single string which is separated by HTML \if{html}{\out{<br>}} tags.
}
\usage{
all_values(x)
}
\arguments{
\item{x}{Named vector for which to concatenate names and values.}
}
\value{
A character vector containing the names and formatted values of the input vector, separated by line breaks.
}
\description{
Concatenate Names and Values of a Named Vector
}
\examples{
my_vector <- c(apples = 5, bananas = 3, oranges = 2)
all_values(my_vector)

}
