% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeStatistics.R
\name{computeStatistics}
\alias{computeStatistics}
\title{Compute Statistics}
\usage{
computeStatistics(df, by_cond = NULL, do_indicator = "drop_out_idx",
  no_of_vars)
}
\description{
Compute Statistics
}

