% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_kpm.R
\name{do_kpm}
\alias{do_kpm}
\title{Kaplan-Meier}
\usage{
do_kpm(d, qs, condition_col, model_fit)
}
\arguments{
\item{d}{dataset}

\item{qs}{character vector of questions column indicator.}

\item{condition_col}{character column denoting the experimental condition}

\item{model_fit}{?}
}
\description{
Kaplan-Meier
}
