% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_kpm.R
\name{do_kpm_plot}
\alias{do_kpm_plot}
\title{Draw a Kaplan Meier Plot}
\usage{
do_kpm_plot(kds, sel_cond_kpm, kpm_ci, color_palette_kp, full_scale_kpm)
}
\arguments{
\item{kds}{description}

\item{sel_cond_kpm}{selected experimental conditions.}

\item{kpm_ci}{confidence bands}

\item{color_palette_kp}{different color palettes}

\item{full_scale_kpm}{?}
}
\description{
Draw a Kaplan Meier Plot
}
