% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_ks.R
\name{do_ks}
\alias{do_ks}
\title{Compute Kolmogorov-Smirnov Test for most extreme conditions}
\usage{
do_ks(data, questions)
}
\arguments{
\item{data}{A dataset made from dropR::compute_stats(), containing information on the percent remaining per question per condition}

\item{questions}{Number of questions to be included in analysis}
}
\description{
Compute Kolmogorov-Smirnov Test for most extreme conditions
}
