% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_or_table.R
\name{do_or_table}
\alias{do_or_table}
\title{Dropout Odds Ration Table}
\usage{
do_or_table(d, chisq_question, sel_cond_chisq)
}
\arguments{
\item{d}{data.frame stats table as computed by compute_stats.}
}
\description{
Dropout Odds Ration Table
}
