% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dropR-package.R
\docType{package}
\name{dropR-package}
\alias{dropR}
\alias{dropR-package}
\title{dropR: Drop out analysis in R}
\description{
Analysis of dropout in surveys and experimental setups.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://iscience-kn.github.io/dropR/}
  \item \url{https://github.com/iscience-kn/dropR}
}

}
\author{
\strong{Maintainer}: Annika Tave Overlander \email{annika-tave.overlander@uni.kn}

Authors:
\itemize{
  \item Matthias Bannert \email{bannert@kof.ethz.ch}
}

Other contributors:
\itemize{
  \item Ulf-Dietrich Reips [contributor]
}

}
\keyword{internal}
