% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dropRdemo.R
\name{dropRdemo}
\alias{dropRdemo}
\title{Demo Dataset for Dropout in an Online Survey}
\format{
A data frame with 246 rows and 54 variables:
\describe{
  \item{obs_id}{Observation ID}
  \item{vi_1}{columns containing variables}
  \item{vi_2}{columns containing variables}
  \item{vi_3}{columns containing variables}
  \item{vi_4}{columns containing variables}
  \item{vi_5}{columns containing variables}
  \item{vi_6}{columns containing variables}
  \item{vi_7}{columns containing variables}
  \item{vi_8}{columns containing variables}
  \item{vi_9}{columns containing variables}
  \item{vi_10}{columns containing variables}
  \item{vi_11}{columns containing variables}
  \item{vi_12}{columns containing variables}
  \item{vi_13}{columns containing variables}
  \item{vi_14}{columns containing variables}
  \item{vi_15}{columns containing variables}
  \item{vi_16}{columns containing variables}
  \item{vi_17}{columns containing variables}
  \item{vi_18}{columns containing variables}
  \item{vi_19}{columns containing variables}
  \item{vi_20}{columns containing variables}
  \item{vi_21}{columns containing variables}
  \item{vi_22}{columns containing variables}
  \item{vi_23}{columns containing variables}
  \item{vi_24}{columns containing variables}
  \item{vi_25}{columns containing variables}
  \item{vi_26}{columns containing variables}
  \item{vi_27}{columns containing variables}
  \item{vi_28}{columns containing variables}
  \item{vi_29}{columns containing variables}
  \item{vi_30}{columns containing variables}
  \item{vi_31}{columns containing variables}
  \item{vi_32}{columns containing variables}
  \item{vi_33}{columns containing variables}
  \item{vi_34}{columns containing variables}
  \item{vi_35}{columns containing variables}
  \item{vi_36}{columns containing variables}
  \item{vi_37}{columns containing variables}
  \item{vi_38}{columns containing variables}
  \item{vi_39}{columns containing variables}
  \item{vi_40}{columns containing variables}
  \item{vi_41}{columns containing variables}
  \item{vi_42}{columns containing variables}
  \item{vi_43}{columns containing variables}
  \item{vi_44}{columns containing variables}
  \item{vi_45}{columns containing variables}
  \item{vi_46}{columns containing variables}
  \item{vi_47}{columns containing variables}
  \item{vi_48}{columns containing variables}
  \item{vi_49}{columns containing variables}
  \item{vi_50}{columns containing variables}
  \item{vi_51}{columns containing variables}
  \item{vi_52}{columns containing variables}
  \item{experimental_condition}{experimental condition}
}
}
\source{
{dropRdemo} dropRdemo R package.
}
\description{
Simulated demo dataset for dropout in a survey.
}
