% Generated by roxygen2 (4.0.2): do not edit by hand
\name{extract_drop_out_from_df}
\alias{extract_drop_out_from_df}
\title{Extract Drop Out from a Data.Frame}
\usage{
extract_drop_out_from_df(df, q_pos)
}
\arguments{
\item{df}{a data.frame}

\item{q_pos}{columns that contain questions}
}
\description{
Find drop in Data.frame that contains multiple
questions that had been asked sequentially.
}
\examples{
data(dropout)
dropout$pos <- extract_drop_out_from_df(dropout,2:10)
dropout$pos
}

