% Generated by roxygen2 (4.0.2): do not edit by hand
\name{find_drop_out}
\alias{find_drop_out}
\title{Find the Position of Drop Out in a Vector}
\usage{
find_drop_out(v, clnms)
}
\arguments{
\item{v}{a vector}

\item{clnms}{specify the parts that actually hold questions by character names.}
}
\description{
Check consecutive NAs from backend of a vector.
}

