% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_drop_out.R
\name{find_drop_out}
\alias{find_drop_out}
\title{Find the Position of Drop Out in a Vector}
\usage{
find_drop_out(v, clnms)
}
\arguments{
\item{v}{a vector}

\item{clnms}{specify the parts that actually hold questions by character names.}
}
\description{
Check consecutive NAs from backend of a vector. 
`r lifecycle::badge("deprecated")`
}
\details{
`find_drop_out()` is deprecated in favor of [add_dropout_idx()], which is the 
basis for dropout analysis using `dropR`.
}
