% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_odds.R
\name{get_odds_ratio}
\alias{get_odds_ratio}
\title{Compute Odds Ratio}
\usage{
get_odds_ratio(a, b)
}
\arguments{
\item{a}{numeric probability value between 0 and 1.}

\item{b}{numeric probability value between 0 and 1.}
}
\description{
Computes and odds ratio given two probabilites.
}
