% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_kpm.R
\name{get_survdiff}
\alias{get_survdiff}
\title{Test Survival Curve Difference}
\usage{
get_survdiff(d, cond, test_type)
}
\arguments{
\item{d}{description}

\item{cond}{description}

\item{test_type}{?}
}
\description{
Test Survival Curve Difference
}
