% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_curve.R
\name{plot_do_curve}
\alias{plot_do_curve}
\title{Plot Dropout Curves}
\usage{
plot_do_curve(
  d,
  linetypes = TRUE,
  stroke_width = 1,
  full_scale = TRUE,
  show_points = FALSE,
  color_palette = "color_blind"
)
}
\arguments{
\item{d}{data.frame containing dropout statistics table computed by compute_stats.
Make sure your stats table contains a do_idx column indexing all items sequentially.}

\item{linetypes}{boolean show we use different linetypes, defaults to TRUE.}

\item{stroke_width}{numeric stroke width, defaults to 1.}

\item{full_scale}{boolean should scale range from 0,100? Defaults to TRUE, FALSE cuts off at max.}

\item{show_points}{boolean should lines show points? Defaults to FALSE.}

\item{color_palette}{character indicating which color palette to use. Defaults to 'color_blind',
alternatively choose 'gray'.}
}
\description{
This functions uses ggplot to create drop out curves.
}
